Updating Makefile...
Added a generic rule for .c files in src/. Don't know how to add a generic
rule to list the library object files though (we don't want the _main code
getting backed into the library). If anybody know how to fix this then go
for it. Otherwise, the list of library object files is still static.